+++ tests/conftest.py
@@ -1,4 +1,5 @@
 import random
+from typing import Callable
 
 import pytest
 import torch
@@ -7,32 +8,103 @@ import cheetah
 from cheetah.utils import is_mps_available_and_functional
 
 ELEMENT_SUBCLASSES_ARGS = {
-    cheetah.Aperture: {},
-    cheetah.BPM: {},
-    cheetah.Cavity: {"length": torch.tensor(1.0)},
-    cheetah.CustomTransferMap: {"predefined_transfer_map": torch.eye(7)},
-    cheetah.Dipole: {"length": torch.tensor(1.0), "angle": torch.tensor([1.0, -2.0])},
-    cheetah.Drift: {"length": torch.tensor([1.0, -1.0])},
+    cheetah.Aperture: {"inactive": {"is_active": False}, "active": {"is_active": True}},
+    cheetah.BPM: {"inactive": {"is_active": False}, "active": {"is_active": True}},
+    cheetah.Cavity: {"default": {"length": torch.tensor(1.0)}},
+    cheetah.CustomTransferMap: {"identity": {"predefined_transfer_map": torch.eye(7)}},
+    cheetah.Dipole: {
+        "linear": {
+            "length": torch.tensor(1.0),
+            "angle": torch.tensor([1.0, -2.0]),
+            "tracking_method": "linear",
+        },
+        "second_order": {
+            "length": torch.tensor(1.0),
+            "angle": torch.tensor([1.0, -2.0]),
+            "tracking_method": "second_order",
+        },
+        "drift_kick_drift": {
+            "length": torch.tensor(1.0),
+            "angle": torch.tensor([1.0, -2.0]),
+            "tracking_method": "drift_kick_drift",
+        },
+    },
+    cheetah.Drift: {
+        "linear": {"length": torch.tensor([1.0, -1.0]), "tracking_method": "linear"},
+        "second_order": {
+            "length": torch.tensor([1.0, -1.0]),
+            "tracking_method": "second_order",
+        },
+        "drift_kick_drift": {
+            "length": torch.tensor([1.0, -1.0]),
+            "tracking_method": "drift_kick_drift",
+        },
+    },
     cheetah.HorizontalCorrector: {
-        "length": torch.tensor(1.0),
-        "angle": torch.tensor([1.0, -2.0]),
+        "default": {"length": torch.tensor(1.0), "angle": torch.tensor([1.0, -2.0])}
     },
-    cheetah.Marker: {},
+    cheetah.Marker: {"default": {}},
     cheetah.Quadrupole: {
-        "length": torch.tensor(1.0),
-        "k1": torch.tensor([1.0, -2.0]),
+        "linear": {
+            "length": torch.tensor(1.0),
+            "k1": torch.tensor([1.0, -2.0]),
+            "tracking_method": "linear",
+        },
+        "second_order": {
+            "length": torch.tensor(1.0),
+            "k1": torch.tensor([1.0, -2.0]),
+            "tracking_method": "second_order",
+        },
+        "drift_kick_drift": {
+            "length": torch.tensor(1.0),
+            "k1": torch.tensor([1.0, -2.0]),
+            "tracking_method": "drift_kick_drift",
+        },
     },
-    cheetah.RBend: {"length": torch.tensor(1.0), "angle": torch.tensor([1.0, -2.0])},
-    cheetah.Screen: {},
-    cheetah.Segment: {"elements": [cheetah.Drift(length=torch.tensor(1.0))]},
-    cheetah.Sextupole: {"length": torch.tensor(1.0), "k2": torch.tensor([1.0, -2.0])},
-    cheetah.Solenoid: {"length": torch.tensor(1.0), "k": torch.tensor([1.0, -2.0])},
-    cheetah.SpaceChargeKick: {"effect_length": torch.tensor(1.0)},
-    cheetah.TransverseDeflectingCavity: {"length": torch.tensor(1.0)},
-    cheetah.Undulator: {"length": torch.tensor(1.0)},
+    cheetah.RBend: {
+        "linear": {
+            "length": torch.tensor(1.0),
+            "angle": torch.tensor([1.0, -2.0]),
+            "tracking_method": "linear",
+        },
+        "second_order": {
+            "length": torch.tensor(1.0),
+            "angle": torch.tensor([1.0, -2.0]),
+            "tracking_method": "second_order",
+        },
+        "drift_kick_drift": {
+            "length": torch.tensor(1.0),
+            "angle": torch.tensor([1.0, -2.0]),
+            "tracking_method": "drift_kick_drift",
+        },
+    },
+    cheetah.Screen: {"default": {}},
+    cheetah.Segment: {
+        "default": {"elements": [cheetah.Drift(length=torch.tensor(1.0))]}
+    },
+    cheetah.Sextupole: {
+        "linear": {
+            "length": torch.tensor(1.0),
+            "k2": torch.tensor([1.0, -2.0]),
+            "tracking_method": "linear",
+        },
+        "second_order": {
+            "length": torch.tensor(1.0),
+            "k2": torch.tensor([1.0, -2.0]),
+            "tracking_method": "second_order",
+        },
+    },
+    cheetah.Solenoid: {
+        "default": {"length": torch.tensor(1.0), "k": torch.tensor([1.0, -2.0])}
+    },
+    cheetah.SpaceChargeKick: {"default": {"effect_length": torch.tensor(1.0)}},
+    cheetah.TransverseDeflectingCavity: {
+        "inactive": {"length": torch.tensor(1.0), "voltage": torch.tensor(0.0)},
+        "active": {"length": torch.tensor(1.0), "voltage": torch.tensor(1e6)},
+    },
+    cheetah.Undulator: {"default": {"length": torch.tensor(1.0)}},
     cheetah.VerticalCorrector: {
-        "length": torch.tensor(1.0),
-        "angle": torch.tensor([1.0, -2.0]),
+        "default": {"length": torch.tensor(1.0), "angle": torch.tensor([1.0, -2.0])}
     },
 }
 
@@ -46,59 +118,28 @@ def pytest_addoption(parser) -> None:
 
 def pytest_generate_tests(metafunc: pytest.Metafunc) -> None:
     """
-    Add a `pytest.mark.parametrize`-like marker that runs a test with an MWE of each
-    Cheetah `Element` subclass.
-
-    This marker can be used by adding the `@pytest.mark.for_every_element` marker to
-    a test function. The user must specify the argument name in the marker via the
-    `arg_name` keyword argument, and define an argument of that name in the test
-    function's signature. Element classes can be excluded from the test by using the
-    `except_for` keyword argument, which takes a list of element classes.
+    Add a `pytest.mark.parametrize`-like marker that runs a test with a test case for
+    each Cheetah `Element` subclass.
     """
-    for_every_mwe_element_marker = metafunc.definition.get_closest_marker(
+    for_every_element_marker = metafunc.definition.get_closest_marker(
         "for_every_element"
     )
 
-    if for_every_mwe_element_marker is None:
+    if for_every_element_marker is None:
         # No marker found, return early
         return
 
     arg_name = (
-        for_every_mwe_element_marker.args[0]
-        if for_every_mwe_element_marker.args[0] is not None
-        else for_every_mwe_element_marker.kwargs["arg_name"]
-    )
-    except_for = (
-        for_every_mwe_element_marker.args[1]
-        if len(for_every_mwe_element_marker.args) > 1
-        else for_every_mwe_element_marker.kwargs.get("except_for", [])
+        for_every_element_marker.args[0]
+        if for_every_element_marker.args[0] is not None
+        else for_every_element_marker.kwargs["arg_name"]
     )
 
-    # Recursively discover all subclasses of `Element`
-    def get_all_subclasses(cls):
-        for subclass in cls.__subclasses__():
-            yield from get_all_subclasses(subclass)
-            yield subclass
-
-    all_element_subclasses = get_all_subclasses(cheetah.Element)
-
-    # Remove subclasses according to `except_for`
-    element_subclasses_to_test = [
-        subclass for subclass in all_element_subclasses if subclass not in except_for
-    ]
-
-    # Generate minimal working examples
-    element_mwe_dict = {
-        subclass: (
-            subclass(**ELEMENT_SUBCLASSES_ARGS[subclass]).clone()
-            if subclass in ELEMENT_SUBCLASSES_ARGS
-            else pytest.param(None, marks=pytest.mark.fail_because_no_mwe_args_defined)
-        )
-        for subclass in element_subclasses_to_test
-    }
-
-    metafunc.parametrize(
-        arg_name, element_mwe_dict.values(), ids=element_mwe_dict.keys()
+    for_every_element(
+        metafunc,
+        arg_name,
+        except_if=for_every_element_marker.kwargs.get("except_if"),
+        xfail_if=for_every_element_marker.kwargs.get("xfail_if"),
     )
 
 
@@ -129,13 +170,13 @@ def seed_random_generators(request):
 
 
 @pytest.fixture(autouse=True)
-def fail_because_no_mwe_args_defined(request):
+def fail_because_no_test_case_defined(request):
     """
-    Mark a test to fail because no MWE args are defined for the given `Element`
+    Mark a test as failing because no test cases are defined for the given `Element`
     subclass.
     """
-    if request.node.get_closest_marker("no_mwe") is not None:
-        pytest.fail("No MWE args are defined for this Element subclass.")
+    if request.node.get_closest_marker("fail_because_no_test_case_defined") is not None:
+        pytest.fail("No test cases are defined for this Element subclass.")
 
 
 @pytest.fixture
@@ -150,3 +191,68 @@ def default_torch_dtype(request):
     yield tmp_dtype_for_test
 
     torch.set_default_dtype(previous_dtype)
+
+
+def for_every_element(
+    metafunc: pytest.Metafunc,
+    arg_name: str,
+    except_if: Callable[[cheetah.Element], bool] | None = None,
+    xfail_if: Callable[[cheetah.Element], bool] | None = None,
+) -> None:
+    """
+    This marker can be used by adding the `@pytest.mark.for_every_element` marker to a
+    test function. The user must specify the argument name in the marker via the first
+    positional argument or the `arg_name` keyword argument, and define an argument of
+    that name in the test function's signature.
+
+    The `except_if` keyword argument provides means to filter which test cases are
+    executed. It expects a single-argument lambda expression that evaluates to a bool
+    and is passed to the `Element` under test one-by-one.
+
+    The `xfail_if` keyword argument can be used to mark test cases are are supposed to
+    fail. It takes a lambda expression with the same signature as `except_if`.
+    """
+    except_if = except_if if except_if is not None else lambda _: False
+    xfail_if = xfail_if if xfail_if is not None else lambda _: False
+
+    # Recursively discover all subclasses of `Element`
+    def get_all_subclasses(cls):
+        for subclass in cls.__subclasses__():
+            yield from get_all_subclasses(subclass)
+            yield subclass
+
+    all_element_subclasses = get_all_subclasses(cheetah.Element)
+
+    # Generate test cases for all element subclasses
+    test_cases = []
+    for subclass in all_element_subclasses:
+        if subclass in ELEMENT_SUBCLASSES_ARGS:
+            subclass_test_cases = ELEMENT_SUBCLASSES_ARGS[subclass]
+            for label, test_case_args in subclass_test_cases.items():
+                # The clone prevents tests from modifying the test cases, which is
+                # especially relevant for `Segment`. This is necessary since the
+                # subclass constructors reference their arguments instead of copying.
+                test_case = subclass(**test_case_args).clone()
+
+                if not except_if(test_case):
+                    test_cases.append(
+                        pytest.param(
+                            test_case,
+                            id=(
+                                f"{subclass.__name__}-{label}"
+                                if len(subclass_test_cases) > 1
+                                else subclass.__name__
+                            ),
+                            marks=pytest.mark.xfail if xfail_if(test_case) else (),
+                        )
+                    )
+        else:
+            test_cases.append(
+                pytest.param(
+                    None,
+                    id=subclass.__name__,
+                    marks=pytest.mark.fail_because_no_test_case_defined,
+                )
+            )
+
+    metafunc.parametrize(arg_name, test_cases)